Link to example config & show how to trigger commands with keybinding
For some reason I could not find the example config even though it was clearly in the repo's root.

By linking it in the config docs its easy to place config parts into context.

It also took me a while to realize I could trigger commands by binding to ':command<Enter>' . Therefore it seemed like a good idea to show that too.